backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
google-generativeai==0.8.3
deepgram-sdk==3.7.7
pymongo[srv]==4.10.1
python-dotenv==1.0.1
pdfplumber==0.11.4
websockets==13.1
pydantic-settings==2.5.2
